[branding] refs #156 Rename trezor.{c,d,o,h} files
