+++ tiny-firmware/README.md
@@ -177,7 +177,7 @@ Here is a quick presentation of most important files of the firmware:
 * firmware/fsm.c: all the messages received in the firmware correspond to a call to a function in fsm.c, the corresponding between messages and function is defined in protob/messages_map.h (generated file)
 * firmware/storage.c: manages the persistent (persistent when pluging out the power supply) memory of the firmware.
 * oled.c/layout.c: manage screen display
-* firmware/trezor.c: main entry point
+* firmware/skywallet.c: main entry point
 
 On bootloader side it is worth mentioning:
