+++ tiny-firmware/firmware/fsm_impl.c
@@ -16,7 +16,7 @@
 #include <libopencm3/stm32/flash.h>
 
 
-#include "trezor.h"
+#include "skywallet.h"
 #include "fsm.h"
 #include "messages.h"
 #include "bip32.h"
