+++ tiny-firmware/firmware/main.c
@@ -9,7 +9,7 @@
  *
  */
 
-#include "trezor.h"
+#include "skywallet.h"
 #include "oled.h"
 #include "bitmaps.h"
 #include "util.h"
